requirements.txt
torch
Pillow
matplotlib
numpy
scikit-learn
